requirements.txt
static-ffmpeg
yt-dlp
requests
azure-storage-blob
python-dotenv
chromadb
openai
pandas
pdfplumber
openpyxl
langchain
langchain-openai
openai
azure-identity
langgraph
langgraph-sdk
langchain-core
langchain-community
orjson
httpx
tenacity
structlog
cloudpickle
biopython